Create Stage4-HelpOnTheWay
Now our game has become quite hard, and your chances of guessing the word depend on the size of the list. In our case with four words, there is only a 25% chance, so let's have mercy on the player and add a hint for them.

Objectives:

As in the previous stage, you should use the following word list: 'python', 'java', 'kotlin', 'javascript'
Show the first 3 letters after the computer chose a word from the list. Hidden letters should be replaced with hyphens ("-").